Remove the `pkg/chain.Handle` interface
This interface was an artificial bag for
more specialized chain interfaces. The client
code rarely needs all of them at once. In order
to make things simpler, we decided to remove the
`chain.Handle` interface and let the client
code depend on specific chain interfaces.
